Moved the SOAP featurizer from matminer.featurizers.structure to matminer.featurizers.site. Also improved the interface and added proper labels.
